+++ packages/ketcher-react/src/script/editor/operations/Text/index.ts
@@ -1,0 +1,3 @@
+export * from './TextCreateDelete'
+export * from './TextUpdate'
+export * from './TextMove'
