+++ packages/ketcher-react/src/script/editor/operations/index.ts
@@ -28,4 +28,4 @@ export * from './rgroup'
 export * from './rxn'
 export * from './simpleObject'
 export * from './sgroup'
-export * from './text'
+export * from './Text'
